9b3d285d:net

+++ net/websockets/websocket_deflate_parameters.h
@@ -5,14 +5,13 @@
 #ifndef NET_WEBSOCKETS_WEBSOCKET_DEFLATE_PARAMETERS_H_
 #define NET_WEBSOCKETS_WEBSOCKET_DEFLATE_PARAMETERS_H_
 
-#include <stdint.h>
-
 #include <string>
 
 #include "base/logging.h"
 #include "net/base/net_export.h"
 #include "net/websockets/websocket_deflater.h"
 #include "net/websockets/websocket_extension.h"
+#include "starboard/types.h"
 
 namespace net {
 
@@ -85,7 +84,8 @@ class NET_EXPORT_PRIVATE WebSocketDeflateParameters {
   // |bits| must be valid as a max_window_bits value.
   void SetServerMaxWindowBits(int bits) {
     DCHECK(IsValidWindowBits(bits));
-    server_max_window_bits_ = WindowBits(bits, true, true);
+    server_max_window_bits_ =
+        WindowBits(static_cast<int16_t>(bits), true, true);
   }
   void SetClientMaxWindowBits() {
     client_max_window_bits_ = WindowBits(0, true, false);
@@ -93,7 +93,8 @@ class NET_EXPORT_PRIVATE WebSocketDeflateParameters {
   // |bits| must be valid as a max_window_bits value.
   void SetClientMaxWindowBits(int bits) {
     DCHECK(IsValidWindowBits(bits));
-    client_max_window_bits_ = WindowBits(bits, true, true);
+    client_max_window_bits_ =
+        WindowBits(static_cast<int16_t>(bits), true, true);
   }
 
   int PermissiveServerMaxWindowBits() const {
